Fix static issues in AudioD
:Release Notes:
Implement to fix static issues in AudioD

:Detailed Notes:
Static issues should be fixed in AudioD

:Testing Performed:
QWTC-6279 QWTC-6273 QWTC-5979 QWTC-6234

:QA Notes:
NA

:Issues Addressed:
[WRO-10600] Analyze and fix the static issues in OSE
[CID-9331506] Overflowed return value
[CID-9331495] Overflowed return value
[CID-9331489] Overflowed return value
[CID-9331480] Overflowed return value

Reviewed-on: http://gpro.lge.com/c/webosose/audiod-pro/+/337871
Tested-by: Velicheti Sita <velicheti.sita@lge.com>

+++ include/PulseAudioMixer.h
@@ -158,7 +158,7 @@ private:
 
     // Function to send message to pulseaudio
     bool sendHeaderToPA(char *data, paudiodMsgHdr audioMsgHdr);
-    template<typename T>bool sendDataToPule (uint8_t msgType, uint8_t msgID, T subObj);
+    template<typename T>bool sendDataToPule (uint32_t msgType, uint32_t msgID, T subObj);
 };
 
 #endif //PULSEAUDIOMIXER_H_
